Add files via upload
First release using data stored in MySQL DB. Also, on messages sent, the DB is updated to store the information about the date when the last message was sent to the recipient address. Dynamic library required for communication with DB was downloaded (from site at URL https://www.dll-files.com/mysql.data.dll.html) and attached to the project (stored in dll folder). The list of the recipients in DB is exported only if the current date is more than 30 days after the last message sent. In case this condition is not met, their e-mail address is not dumped from DB to file, which is used for looping in the email addresses for sending.