Add files via upload
The email body content is created in a professional application, generating newsletter's body. Just go and edit your values. Images are stored in your webserver, so the message is not eating a valued free space in your mailbox. In the fresh dumped SQL table are records I found on Internet. The value of 0 means: do not use the message, and value 1 means no issue was recognized for this address. On the run of the script, each address is updated with the date of last message sent, not to flood the recipients too often.